ml-service/pyproject.toml
[project]
name = "ml-service"
version = "1.0.0"
description = "ML Service for Pawn AI - Gold Price Prediction and Risk Assessment"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "numpy>=2.2.0",
    "pandas>=2.2.0",
    "prophet>=1.1.6",
    "scikit-learn>=1.6.0",
    "xgboost>=2.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "joblib>=1.5.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
